requirements.txt
aiohttp>=3.9.0
aiogram>=3.3.0
python-dotenv==1.0.1 